shadowing_app/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
openai>=1.30.0
google-genai>=1.0.0
python-multipart>=0.0.9
pydub>=0.25.1
